Add bindings for Java i18n classes (#6711)
