+++ framework/src/play/src/main/scala/play/api/i18n/I18nModule.scala
@@ -1,0 +1,31 @@
+/*
+ * Copyright (C) 2009-2016 Lightbend Inc. <https://www.lightbend.com>
+ */
+package play.api.i18n
+
+import play.api.{ Configuration, Environment }
+import play.api.inject.Module
+
+class I18nModule extends Module {
+  def bindings(environment: Environment, configuration: Configuration) = {
+    Seq(
+      bind[Langs].to[DefaultLangs],
+      bind[MessagesApi].to[DefaultMessagesApi],
+      bind[play.i18n.MessagesApi].toSelf,
+      bind[play.i18n.Langs].toSelf
+    )
+  }
+}
+
+/**
+ * Injection helper for i18n components
+ */
+trait I18nComponents {
+
+  def environment: Environment
+  def configuration: Configuration
+
+  lazy val messagesApi: MessagesApi = new DefaultMessagesApi(environment, configuration, langs)
+  lazy val langs: Langs = new DefaultLangs(configuration)
+
+}
